// app/lib/extensions/interfaces.ts
export interface MergeSettings {
    /**
     * Root of source manifests
     */
    root: string;
    
    /**
     * List of globs for searching for partial manifests
     */
    manifestGlobs: string[];
    
    /**
     * Highest priority partial manifest
     */
    overrides: any;
    
    /**
     * True to bypass validation during packaging.
     */
    bypassValidation: boolean;
}

export interface PackageSettings {
    /**
     * Path to the generated vsix
     */
    outputPath: string;
    
    /**
     * Path to the root of localized resource files
     */
    locRoot: string;
    
}